Aggregate instances of cultural properties in h3 tiles
