+++ ui/src/features/quotes/Quotes.module.css
