+++ ui/src/utils/getSteppedResolutionZoomPair.ts
